requirements.txt
alembic==1.16.4
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
greenlet==3.2.4
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
narwhals==2.1.1
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
plotly==6.3.0
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
SciencePlots==2.1.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.43
sympy==1.13.3
tensorboardX==2.6.4
threadpoolctl==3.6.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
yacs==0.1.8
